requirements.txt
appnope==0.1.0
backcall==0.1.0
confusable-homoglyphs==3.1.1
decorator==4.3.0
Django==2.0.5
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.2.0
django-registration==2.4.1
django-sass-processor==0.7
djangorestframework==3.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
libsass==0.14.5
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.4
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
